refactor(planning): extract the mobile summary/allocation cards + section shell (#467)
Second planning-views slice. Move MobilePlanningView's presentational cards and
states into a new `planningCards.tsx`: SalaryCard, AllocationSummaryCard (+ its
StackedBar), the NoPlanState / PlanErrorState empty states, the collapsible
BudgetSection wrapper, and FixedExpIcon. All are module-level and purely prop-driven
— the view still owns the data + handlers and passes them in. No behavior or markup
change; now-unused lucide imports (Wallet / ChevronDown / ChevronUp / Calendar /
RefreshCw) drop out of the view.

Parity: the PlanningClient / MobilePlanningView component tests + the planning E2E
(create plan — which renders the salary + allocation cards and the budget sections).

Verified: full unit suite 1261 passed; typecheck 0 errors; lint 0 errors;
git diff --check clean; E2E planning (3) passed. MobilePlanningView 1216 → 910 lines.
